[BUG FIX] Fix file transfer bug
Bug:
If offset is set when transferring file, offset mismatch exception will occur.

Cause:
The offset in request will always begin with zero because the offset passed to RemoteManager.readFile is to seek to in RemoteFileWriter(with LocalFileSource) and RemoteFileReader(with StreamSink). However, the StreamSink's read() will check the offset passed in with the real offset.

Fix:
Pass the real offset to StreamSink's read().